# EVM: Move from an EnumSet for halt reason to a simple Optional.