--- a/ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/results/DebugTraceTransactionResult.java
+++ b/ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/results/DebugTraceTransactionResult.java
@@ -48,9 +48,10 @@ public class DebugTraceTransactionResult {
   }
 
   private static StructLog createStructLog(final TraceFrame frame) {
-    return frame.getExceptionalHaltReasons().isEmpty()
-        ? new StructLog(frame)
-        : new StructLogWithError(frame);
+    return frame
+        .getExceptionalHaltReason()
+        .map(__ -> (StructLog) new StructLogWithError(frame))
+        .orElse(new StructLog(frame));
   }
 
   @JsonGetter(value = "structLogs")