--- a/ethereum/core/src/main/java/org/hyperledger/besu/ethereum/vm/operations/JumpOperation.java
+++ b/ethereum/core/src/main/java/org/hyperledger/besu/ethereum/vm/operations/JumpOperation.java
@@ -22,7 +22,6 @@ import org.hyperledger.besu.ethereum.vm.ExceptionalHaltReason;
 import org.hyperledger.besu.ethereum.vm.GasCalculator;
 import org.hyperledger.besu.ethereum.vm.MessageFrame;
 
-import java.util.EnumSet;
 import java.util.Optional;
 
 import org.apache.tuweni.units.bigints.UInt256;
@@ -46,9 +45,7 @@ public class JumpOperation extends AbstractOperation {
 
   @Override
   public Optional<ExceptionalHaltReason> exceptionalHaltCondition(
-      final MessageFrame frame,
-      final EnumSet<ExceptionalHaltReason> previousReasons,
-      final EVM evm) {
+      final MessageFrame frame, final EVM evm) {
     final Code code = frame.getCode();
 
     final UInt256 potentialJumpDestination = UInt256.fromBytes(frame.getStackItem(0));