… ethereum/api/src/test/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/DebugTraceBlockTest.java
… ethereum/api/src/test/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/DebugTraceTransactionTest.java
… ethereum/core/src/main/java/org/hyperledger/besu/ethereum/vm/ehalt/ExceptionalHaltManager.java
… ethereum/core/src/test/java/org/hyperledger/besu/ethereum/vm/DebugOperationTracerTest.java
… ethereum/core/src/test/java/org/hyperledger/besu/ethereum/vm/operations/JumpOperationTest.java
… ethereum/core/src/test/java/org/hyperledger/besu/ethereum/vm/operations/JumpSubOperationTest.java
… ethereum/core/src/test/java/org/hyperledger/besu/ethereum/vm/operations/ReturnSubOperationTest.java
… ethereum/core/src/test/java/org/hyperledger/besu/ethereum/vm/operations/SStoreOperationTest.java