Internally we used to use an enum set to track halt reasons and we would
track multiple halt reasons.  However, what the halt reason is does not
matter to reference tests and tracing, only that a halt occurred.

Replace the EnumSet with an Optional and trace only one revert reason.
This saves us time in enumset management and also allows us to return
quicker once any halt is detected. 

Signed-off-by: Danno Ferrin <danno.ferrin@gmail.com>