@shemnon 2020-06-16
Performance improvement in a tight evm loop is abour 11-14%

Caliper simple open - 199µs -> 170s
Caliper simple transfer - 309µs -> 274 µs